requirements.txt
# IF USING THIRD-PARTY PACKAGES, USE A NEW ENV:
python-dotenv